Separating reading of .enso-source* files into own class
